Fix winrate calculations and minor bugs (#23)
* Fix pairwise-all to get pair match list

As well as make_match function, make_match_all_pairs should
return the pairwise list.

* Fix winrate matrix calcuration

In winrate computation, we have a couple of dataframes for beats and
loss. Even though it depends on how much generated judgments included
in the results, the index of matrix may not be symetric.

e.g. run all pairwise judgment, the last model can be presented only in
opponent columns.

Here we are now thiking about the caluculation of winrate. We have an
addition of a_ij (beats sore) and inverse matrix of b_ij (loss score)
matrix. If we have a symetric index for both of a_ij and b_ij, there
is no matters. However, if the indexes are asynmetrical, "+" operator
will fill with NaN to the element, whose index never appeared in either
a_ij or b_ij. Then, it causes NaN values for the winrate calculation
result.

This commit changes the operator from "+" to ".add" to fill with 0
when the matrix are asynmetrical each other.

* Fix multiprocessing import error

Trivially, "multiprocessing" moudle is in python
standard library instead of "multiprocess".

ref: https://docs.python.org/ja/3/library/multiprocessing.html

* Remove unused StandingsRegistry import

that causes import errors within the current dependencies.